requirements.txt
asgiref==3.2.7
Django==3.0.6
django-filter==2.2.0
django-widget-tweaks==1.4.8
pytz==2020.1
sqlparse==0.3.1
